refactored the card
