--- client/src/Components/CountryCard/Card.css
@@ -1,35 +1,0 @@
-.Card {
-  display: flex;
-  flex-flow: column;
-  justify-content: space-around;
-  height: 300px;
-  width: 100%;
-}
-.Card .Card_Header {
-  display: flex;
-  justify-content: space-between;
-  padding-right: 0.2rem;
-  align-items: center;
-  border: 1px solid black;
-  width: 100%;
-  background: beige;
-}
-.Card .Card_Header .Header_Country-Name {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-  align-items: center;
-  font-weight: 700;
-  font-size: 1.2rem;
-}
-.Card .Card_Header .Header_Country-Name span {
-  margin-right: 0.5rem;
-}
-.Card .Card_Header .Header_Country-Name .closeCountryPanelIcon {
-  /* css for this */
-  border: 3px solid red;
-}
-.Card .Card_Header .Header_Country-Name .closeCountryPanelIcon:hover {
-  cursor: pointer;
-  color: red;
-}
