--- client/src/Components/CountryCard/Card.less
@@ -1,35 +1,0 @@
-.Card {
-  display: flex;
-  flex-flow: column;
-  justify-content: space-around;
-  height: 300px;
-  width: 100%;
-  .Card_Header {
-    display: flex;
-    justify-content: space-between;
-    padding-right: 0.2rem;
-    align-items: center;
-    border: 1px solid black;
-    width: 100%;
-    background: beige;
-    .Header_Country-Name {
-      display: flex;
-      justify-content: center;
-      width: 100%;
-      align-items: center;
-      font-weight: 700;
-      font-size: 1.2rem;
-      span {
-        margin-right: 0.5rem;
-      }
-      .closeCountryPanelIcon {
-        /* css for this */
-        border: 3px solid red;
-        &:hover {
-          cursor: pointer;
-          color: red;
-        }
-      }
-    }
-  }
-}
