--- client/src/Components/CountryCard/Card.test.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import Card from './Card';
-
-it.skip('renders without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(<Card />, div);
-  ReactDOM.unmountComponentAtNode(div);
-});
