--- client/src/Components/CountryCard/CountryPanel.css
@@ -1,19 +1,0 @@
-.CountryPanel {
-  position: fixed;
-  background-color: rgba(255, 255, 255, 0.1);
-  bottom: 0px;
-  left: 0px;
-  z-index: 11;
-  height: auto;
-  width: auto;
-  padding: 10px;
-  animation: Settings 0.5s all;
-  transition-property: all;
-  transition-duration: 0.8s;
-}
-.CountryPanel .CountryPanel-closed {
-  right: -300px;
-}
-.CountryPanel .CountryPanel-open {
-  right: 0px;
-}
