+++ client/src/Components/CountryPanel/CountryPanel.less
@@ -1,0 +1,50 @@
+.CountryPanel {
+  position: fixed;
+  background-color: rgba(255, 255, 255, 0.1);
+  bottom: 0px;
+  left: 0px;
+  z-index: 11;
+  height: auto;
+  width: auto;
+  padding: 10px;
+
+  animation: Settings 0.5s all;
+  transition-property: all;
+  transition-duration: 0.8s;
+
+  .Card {
+    display: flex;
+    flex-flow: column;
+    justify-content: space-around;
+    height: 300px;
+    width: 100%;
+    .Card_Header {
+      display: flex;
+      justify-content: space-between;
+      padding-right: 0.2rem;
+      align-items: center;
+      border: 1px solid black;
+      width: 100%;
+      background: beige;
+      .Header_Country-Name {
+        display: flex;
+        justify-content: center;
+        width: 100%;
+        align-items: center;
+        font-weight: 700;
+        font-size: 1.2rem;
+        span {
+          margin-right: 0.5rem;
+        }
+        .closeCountryPanelIcon {
+          /* css for this */
+          border: 3px solid red;
+          &:hover {
+            cursor: pointer;
+            color: red;
+          }
+        }
+      }
+    }
+  }
+}
